Recognise a couple more walkable plants
Allow mobs to walk through poison ivy seedlings and vines (from the
plantlife and vines modules respectively). This might seem like a
trivial thing to do, but in fact not allowing this causes serious
problems.

Example 1: a vine grows down into the node where a chicken is. Without
this change, the chicken detects this as a collision and in some
circumstances triggers an explosion of failed attempts to find a new
safe location.

Example 2: a poison ivy seedling grows up under the middle of a cow. The
cow is then permanently stuck (and also taxing the server by
continuously trying to get out of the stuck situation).